# back/central-reserve/shared/docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3050
info:
  contact:
    email: backend@example.com
    name: Equipo de Backend
  description: Servicio REST para la gestión de reservas multi-restaurante.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://ejemplo.com/terminos
  title: Restaurant Reservation API
  version: "1.0"
paths:
...
  /businesses:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de todos los negocios del sistema
      parameters:
      - description: Número de página (por defecto 1)
        in: query
        name: page
        type: integer
      - description: Elementos por página (por defecto 10)
        in: query
        name: per_page
        type: integer
      - description: Filtrar por nombre de negocio
        in: query
        name: name
        type: string
      - description: Filtrar por tipo de negocio
        in: query
        name: business_type_id
        type: integer
      - description: Filtrar por estado activo/inactivo
        in: query
        name: is_active
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Negocios obtenidos exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Solicitud inválida
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Token de acceso requerido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener lista de negocios
      tags:
      - businesses
    post:
      consumes:
      - multipart/form-data
      description: Crea un nuevo negocio en el sistema
      parameters:
      - description: Nombre del negocio
        in: formData
        name: name
        required: true
        type: string
      - description: Código del negocio
        in: formData
        name: code
        required: true
        type: string
      - description: ID del tipo de negocio
        in: formData
        name: business_type_id
        required: true
        type: integer
      - description: Zona horaria
        in: formData
        name: timezone
        type: string
      - description: Dirección
        in: formData
        name: address
        type: string
      - description: Descripción
        in: formData
        name: description
        type: string
      - description: Logo del negocio
        in: formData
        name: logo_url
        type: file
      - description: Color primario
        in: formData
        name: primary_color
        type: string
      - description: Color secundario
        in: formData
        name: secondary_color
        type: string
      - description: Dominio personalizado
        in: formData
        name: custom_domain
        type: string
      - description: Habilitar delivery
        in: formData
        name: enable_delivery
        type: boolean
      - description: Habilitar pickup
        in: formData
        name: enable_pickup
        type: boolean
      - description: Habilitar reservas
        in: formData
        name: enable_reservations
        type: boolean
      produces:
      - application/json
      responses:
        "201":
          description: Negocio creado exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Solicitud inválida
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Token de acceso requerido
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear un nuevo negocio
      tags:
      - businesses
...
  /businesses/configured-resources:
    get:
      consumes:
      - application/json
      description: Obtiene una lista paginada de businesses con sus recursos configurados
        y su estado (activo/inactivo)
      parameters:
      - default: 1
        description: Número de página
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        description: Elementos por página
        in: query
        maximum: 100
        minimum: 1
        name: per_page
        type: integer
      - description: Filtrar por ID de business
        in: query
        name: business_id
        type: integer
      - description: Filtrar por ID de tipo de business
        in: query
        name: business_type_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Lista de businesses con recursos obtenida exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Parámetros de consulta inválidos
          schema:
            additionalProperties: true
            type: object
        "401":
          description: No autorizado
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Sin permisos
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno del servidor
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Obtener configuración de recursos de businesses
      tags:
      - Business Resources
...
securityDefinitions:
  BearerAuth:
    description: Token de business JWT con el prefijo **Bearer** (para todos los endpoints)
    in: header
    name: Authorization
    type: apiKey
  BusinessTokenAuth:
    description: Token principal JWT con el prefijo **Bearer** (solo para /auth/business-token)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"